+++ tests/test_internals/test_nwb_extractors.py
@@ -102,19 +102,10 @@ class TestNwbImagingExtractor(unittest.TestCase):
         expected_frames = self.video[frame_idxs, ...]
         np.testing.assert_array_almost_equal(frames_with_array, expected_frames)
 
-        # Test spikeinterface-like behavior for get_video
-        one_element_video_shape = nwb_imaging_extractor.get_video(start_frame=0, end_frame=1, channel=0).shape
-        expected_shape = (1, image_size[0], image_size[1])
-        assert one_element_video_shape == expected_shape
-
         video = nwb_imaging_extractor.get_video()
         expected_video = self.video
 
         np.testing.assert_array_almost_equal(video, expected_video)
-
-    def test_get_frames_indexing_with_single_channel(self):
-        nwb_imaging_extractor = NwbImagingExtractor(file_path=self.file_path)
-        assert_get_frames_return_shape(imaging_extractor=nwb_imaging_extractor)
 
 
 if __name__ == "__main__":