+++ CHANGELOG.md
@@ -4,7 +4,10 @@
 
 ### Fixes
 
-## Improvements
+### Deprecations
+* The 'channel' parameter in get_frames() and get_video() methods is deprecated and will be removed in August 2025.
+
+### Improvements
 * Use `pyproject.toml` for project metadata and installation requirements [#382](https://github.com/catalystneuro/roiextractors/pull/382)